# ports/http/generated/service.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.4 DO NOT EDIT.
package generated

// User defines model for User.
type User struct {
	Address     string `json:"address"`
	FirstName   string `json:"first_name"`
	Id          int64  `json:"id"`
	LastName    string `json:"last_name"`
	PhoneNumber string `json:"phone_number"`
}

// UserCreateParams defines model for UserCreateParams.
type UserCreateParams struct {
	Address     string `json:"address"`
	FirstName   string `json:"first_name"`
	LastName    string `json:"last_name"`
	PhoneNumber string `json:"phone_number"`
}

// UserUpdateParams defines model for UserUpdateParams.
type UserUpdateParams struct {
	Address     *string `json:"address,omitempty"`
	FirstName   *string `json:"first_name,omitempty"`
	LastName    *string `json:"last_name,omitempty"`
	PhoneNumber *string `json:"phone_number,omitempty"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = UserCreateParams

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = UserUpdateParams